+++ CHANGELOG.md
@@ -23,6 +23,10 @@ If you like this project and find it useful, please consider giving it a star on
 
 ## [1.0.0] - Dev branch
 
+### Added
+
+- [bun]: Add Bun version, install, binary, cache, and global module information to `--info`, including the related environment variable values.
+
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
 
 ## [0.0.5] - 2026-07-12